It now compiles without warnings
Yes, I'm a lunatic

+++ src/MS5837.cpp
@@ -47,7 +47,7 @@ bool LANDSHARKS_MS5837::init(TwoWire &wirePort) {
 		_i2cPort->write(MS5837_PROM_READ+i*2);
 		_i2cPort->endTransmission();
 
-		_i2cPort->requestFrom(MS5837_ADDR,2);
+		_i2cPort->requestFrom(MS5837_ADDR,(byte)2);
 		C[i] = (_i2cPort->read() << 8) | _i2cPort->read();
 	}
 
@@ -124,7 +124,7 @@ void LANDSHARKS_MS5837::read() {
 			return;
 		}
 
-		_i2cPort->requestFrom(MS5837_ADDR,3);
+		_i2cPort->requestFrom(MS5837_ADDR,(byte)3);
 		D1_pres = 0;
 		D1_pres = _i2cPort->read();
 		D1_pres = (D1_pres << 8) | _i2cPort->read();
@@ -155,7 +155,7 @@ void LANDSHARKS_MS5837::read() {
 			return;
 		}
 
-		_i2cPort->requestFrom(MS5837_ADDR,3);
+		_i2cPort->requestFrom(MS5837_ADDR,(byte)3);
 		D2_temp = 0;
 		D2_temp = _i2cPort->read();
 		D2_temp = (D2_temp << 8) | _i2cPort->read();
